+++ src/managers/internal/ToastManager/ToastManager.ts
@@ -1,4 +1,5 @@
 import isEqual from 'lodash.isequal';
+import { DEFAULT_TOAST_LIEFTIME } from 'constants/transactions.constants';
 import { UITagsEnum } from 'constants/UITags.enum';
 import { MvxToastList } from 'lib/sdkDappUi';
 import { NotificationsFeedManager } from 'managers/NotificationsFeedManager/NotificationsFeedManager';
@@ -8,7 +9,8 @@ import {
   removeAllCustomToasts,
   removeCustomToast,
   removeTransactionToast,
-  addTransactionToast
+  addTransactionToast,
+  createCustomToast
 } from 'store/actions/toasts/toastsActions';
 import {
   getIsTransactionFailed,
@@ -20,7 +22,7 @@ import { CustomToastType } from 'store/slices/toast/toastSlice.types';
 import { getStore } from 'store/store';
 import { IEventBus } from 'types/manager.types';
 import { ProviderErrorsEnum } from 'types/provider.types';
-import { createUIElement } from 'utils/createUIElement';
+import { ComponentFactory } from 'utils/ComponentFactory';
 import { createToastsFromTransactions } from './helpers/createToastsFromTransactions';
 import { LifetimeManager } from './helpers/LifetimeManager';
 import { ITransactionToast, ToastEventsEnum } from './types';
@@ -30,7 +32,7 @@ interface IToastManager {
 }
 
 export class ToastManager {
-  private lifetimeManager: LifetimeManager;
+  private readonly lifetimeManager: LifetimeManager;
   private isCreatingElement = false;
   private static instance: ToastManager;
   private toastsElement: MvxToastList | null = null;
@@ -38,7 +40,7 @@ export class ToastManager {
   private customToasts: CustomToastType[] = [];
   private successfulToastLifetime?: number;
   private storeToastsSubscription: () => void = () => null;
-  private notificationsFeedManager: NotificationsFeedManager;
+  private readonly notificationsFeedManager: NotificationsFeedManager;
   private eventBusUnsubscribeFunctions: (() => void)[] = [];
   private eventBus: IEventBus<ITransactionToast[] | CustomToastType[]> | null =
     null;
@@ -52,13 +54,16 @@ export class ToastManager {
     this.notificationsFeedManager = NotificationsFeedManager.getInstance();
   }
 
-  public async init({ successfulToastLifetime }: IToastManager = {}) {
+  public async init({
+    successfulToastLifetime = DEFAULT_TOAST_LIEFTIME
+  }: IToastManager = {}) {
     this.successfulToastLifetime = successfulToastLifetime;
 
     this.lifetimeManager.init({ successfulToastLifetime });
 
     this.updateTransactionToastsList();
     this.updateCustomToastList();
+
     await this.subscribeToEventBusNotifications();
 
     this.storeToastsSubscription = this.store.subscribe(
@@ -112,14 +117,24 @@ export class ToastManager {
     return isCompleted;
   }
 
-  public createTransactionToast(toastId: string, totalDuration: number) {
-    addTransactionToast({
+  public createTransactionToast(
+    toastId: string,
+    totalDuration: number
+  ): string {
+    const newToastId = addTransactionToast({
       toastId,
       totalDuration
     });
 
     this.handleCompletedTransaction(toastId);
     this.updateTransactionToastsList();
+    return newToastId;
+  }
+
+  public createCustomToast(toast: CustomToastType): string {
+    const toastId = createCustomToast(toast);
+    this.updateCustomToastList();
+    return toastId;
   }
 
   private async updateTransactionToastsList() {
@@ -185,7 +200,7 @@ export class ToastManager {
     if (!this.isCreatingElement) {
       this.isCreatingElement = true;
 
-      this.toastsElement = await createUIElement<MvxToastList>({
+      this.toastsElement = await ComponentFactory.create<MvxToastList>({
         name: UITagsEnum.TOAST_LIST
       });
 
@@ -205,6 +220,7 @@ export class ToastManager {
 
   private async subscribeToEventBusNotifications() {
     const toastsElement = await this.createToastListElement();
+
     if (!toastsElement) {
       return;
     }