+++ src/methods/transactions/getFailedTransactions.ts
@@ -1,0 +1,6 @@
+import { getState } from 'store/store';
+import { failedTransactionsSelector } from '../../store/selectors/transactionsSelector';
+
+export function getFailedTransactions(state = getState()) {
+  return failedTransactionsSelector(state);
+}