+++ src/methods/transactions/getFailedTransactionsSessions.ts
@@ -1,0 +1,6 @@
+import { getState } from 'store/store';
+import { failedTransactionsSessionsSelector } from '../../store/selectors/transactionsSelector';
+
+export function getFailedTransactionsSessions(state = getState()) {
+  return failedTransactionsSessionsSelector(state);
+}