+++ src/store/actions/transactions/transactionsActions.ts
@@ -9,15 +9,19 @@ import {
   SignedTransactionType
 } from 'types/transactions.types';
 
-export const createTransactionsSession = ({
-  transactions,
-  transactionsDisplayInfo,
-  status
-}: {
+export type CreateTransactionsSessionType = {
   transactions: SignedTransactionType[];
   transactionsDisplayInfo?: TransactionsDisplayInfoType;
   status: TransactionBatchStatusesEnum | TransactionServerStatusesEnum;
-}) => {
+  sessionInformation?: any;
+};
+
+export const createTransactionsSession = ({
+  transactions,
+  transactionsDisplayInfo,
+  status,
+  sessionInformation
+}: CreateTransactionsSessionType) => {
   const sessionId = Date.now().toString();
   getStore().setState(
     ({ transactions: state }) => {
@@ -25,7 +29,8 @@ export const createTransactionsSession = ({
         transactions,
         status,
         transactionsDisplayInfo,
-        interpretedTransactions: {}
+        interpretedTransactions: {},
+        sessionInformation
       };
     },
     false,
@@ -34,7 +39,7 @@ export const createTransactionsSession = ({
   return sessionId;
 };
 
-export const updateTransactionsSession = ({
+export const updateSessionStatus = ({
   sessionId,
   status,
   errorMessage
@@ -59,7 +64,8 @@ export const updateTransactionStatus = ({
 }: {
   sessionId: string;
   transaction: SignedTransactionType;
-}) => {
+}): TransactionBatchStatusesEnum | null => {
+  let newStatus: TransactionBatchStatusesEnum | null = null;
   getStore().setState(
     ({ transactions: state }) => {
       const transactions = state[sessionId]?.transactions;
@@ -74,15 +80,24 @@ export const updateTransactionStatus = ({
           return transaction;
         });
 
-        const status = getTransactionsSessionStatus(transactions);
+        const status = getTransactionsSessionStatus([
+          ...state[sessionId].transactions // Create a copy of the transactions array to avoid Proxy issues
+        ]);
+
         if (status) {
-          state[sessionId].status = status;
+          updateSessionStatus({
+            sessionId,
+            status
+          });
+          newStatus = status;
         }
       }
     },
     false,
     'updateTransactionStatus'
   );
+
+  return newStatus;
 };
 
 export const clearCompletedTransactions = () => {