+++ src/types/theme.types.ts
@@ -1,0 +1,4 @@
+export enum ThemesEnum {
+  light = 'mvx:light-theme',
+  dark = 'mvx:dark-theme'
+}